add shared memory back && extract skipping method out in array mode (0-activated problem remained)

+++ src/main/scala/edu/ustc/nodb/PregelGPU/algorithm/SSSP/GPUController.scala
@@ -161,7 +161,7 @@ class GPUController(vertexSum: Long,
       var invalidDetector = 0.0
       for(j <- sourceList.indices) {
         invalidDetector = resultAttr(i * sourceSize + j)
-        if(invalidDetector < Double.MaxValue) {
+        if(invalidDetector < Int.MaxValue) {
           tempVertexAttr.+=((sourceList(j), resultAttr(i * sourceSize + j)))
         }
       }
